Fix | Replace Listener For Terminatable Middleware
