+++ .gitignore
@@ -12,3 +12,4 @@ testbench.yaml
 vendor
 node_modules
 .php-cs-fixer.cache
+.phpunit.cache/
