--- src/EventServiceProvider.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace Bilfeldt\RequestLogger;
-
-use Bilfeldt\RequestLogger\Listeners\LogRequest;
-use Illuminate\Foundation\Http\Events\RequestHandled;
-use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
-
-class EventServiceProvider extends ServiceProvider
-{
-    protected $listen = [
-        RequestHandled::class => [
-            LogRequest::class,
-        ],
-    ];
-
-    /**
-     * Register any events for your application.
-     *
-     * @return void
-     */
-    public function boot()
-    {
-        parent::boot();
-    }
-}
